// part9/Patientor_Typing_an_Expess_app/ReactWithTypes/src/types.tsx
interface CoursePartBase {
  name: string;
  exerciseCount: number;
}
interface CoursePartBaseWithDescription extends CoursePartBase {
  description: string;
}

interface CoursePartBasic extends CoursePartBaseWithDescription {
  kind: "basic";
}

interface CoursePartGroup extends CoursePartBase {
  groupProjectCount: number;
  kind: "group";
}

interface CoursePartBackground extends CoursePartBaseWithDescription {
  backgroundMaterial: string;
  kind: "background";
}
interface CoursePartSpecial extends CoursePartBaseWithDescription {
  requirements: string[];
  kind: "requirements";
}

export type CoursePart =
  | CoursePartBasic
  | CoursePartGroup
  | CoursePartBackground
  | CoursePartSpecial;